Rename RatisRouting as HistoryManagementRouting
