+++ src/main/kotlin/com/github/davenury/ucac/common/InMemoryHistoryManagement.kt
@@ -7,38 +7,34 @@ import org.slf4j.LoggerFactory
 class InMemoryHistoryManagement(
     private val consensusProtocol: ConsensusProtocol<Change, History>
 ) : HistoryManagement(consensusProtocol) {
-    private val historyStorage: MutableList<ChangeWithAcceptNum> = mutableListOf()
 
     // TODO: think about what's better - if change asks consensus protocol if it
     // can be done or if something higher asks and then calls change
     override suspend fun change(change: Change, acceptNum: Int?): HistoryChangeResult =
-            consensusProtocol.proposeChange(change,acceptNum)
-                .let {
-                    when (it) {
-                        ConsensusFailure -> {
-                            HistoryChangeResult.HistoryChangeFailure
-                        }
-                        ConsensusSuccess -> {
-                            historyStorage.add(ChangeWithAcceptNum(change, acceptNum))
-                            HistoryChangeResult.HistoryChangeSuccess
-                        }
-                        ConsensusResultUnknown -> {
-                            historyStorage.add(ChangeWithAcceptNum(change, acceptNum))
-                            HistoryChangeResult.HistoryChangeSuccess
-                        }
+        consensusProtocol.proposeChange(change, acceptNum)
+            .let {
+                when (it) {
+                    ConsensusFailure -> {
+                        HistoryChangeResult.HistoryChangeFailure
+                    }
+                    ConsensusSuccess -> {
+                        HistoryChangeResult.HistoryChangeSuccess
+                    }
+                    ConsensusResultUnknown -> {
+                        HistoryChangeResult.HistoryChangeSuccess
                     }
                 }
+            }
 
     override fun getLastChange(): ChangeWithAcceptNum? =
         try {
-            historyStorage.last()
+            consensusProtocol.getState()?.last()
         } catch (ex: java.util.NoSuchElementException) {
             logger.error("History is empty!")
             null
         }
 
-    override fun getState(): History? =
-        this.historyStorage
+    override fun getState(): History = consensusProtocol.getState() ?: mutableListOf()
 
     override fun canBeBuild(newChange: Change): Boolean = true
